refactor(app) - rename shared component folder
