+++ src/constants/images.constants.ts
@@ -2,7 +2,7 @@ import logo from '/public/images/logo.png'
 import contactFormImage from '/public/images/contact-form.svg'
 
 export const images = {
-  shared: {
+  common: {
     logo,
   },
